+++ srcs/ClassSchedule.UI.Shared/ViewModels/ShellViewModel.cs
@@ -1,0 +1,17 @@
+using CommunityToolkit.Mvvm.ComponentModel;
+
+namespace ClassSchedule.UI.Shared.ViewModels;
+
+/// <summary>
+/// 壳视图模型, 承载整个应用的视图模型, 负责管理全局状态与导航
+/// </summary>
+public sealed partial class ShellViewModel : ObservableObject
+{
+    /// <summary>
+    /// 构造函数
+    /// </summary>
+    public ShellViewModel()
+    {
+
+    }
+}